+++ bootloader/src/bootloader.c
@@ -31,7 +31,7 @@ SOFTWARE.
 #include "ch.h"
 #include "hal.h"
 
-#define BOOTLOADER_VERSION  0x00010102  /* Version 1.1.2 */
+#define BOOTLOADER_VERSION  0x00010200  /* Version 1.2.0 */
 
 static bootloader_state_t state = BOOTLOADER_STATE_IDLE;
 static systime_t timeout_start = 0;